migrate system job days validation to new mechanism
